seed/pyproject.toml
[tool.poetry]
name = "lucent-etl"
version = "0.1.0"
description = "Data ETL Pipeline"
authors = ["Jacob Brown <jdb897@uowmail.edu.au>"]
package-mode=false

[tool.poetry.dependencies]
python = "^3.12"
boto3 = "^1.40.12"
pandas = "^2.0.0"
psycopg2-binary = "^2.9.0"
openpyxl = "^3.1.0"
langchain = "^0.3.27"
langchain-openai = "^0.3.33"
langchain-postgres = "^0.0.15"
langchain-text-splitters = "^0.3.11"
langchain-community = "^0.3.29"
langgraph = "^0.6.7"
sqlalchemy = "^2.0.43"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
